Made "lca" and compareTo take O(P) time
This changes the "lca" and compareTo methods so that they take O(P) time, where P is the length of the path between the two nodes.

I didn't want to change these methods to take O(P) time if this would significantly affect the performance in use cases where P is large on average (e.g. for a random or worst-case selection of inputs). For example, I didn't want to use a HashSet or even allocate an array. But I finally thought of a way of accomplishing this. "lca" and compareTo are basically doing the same operations as they were before, just in a different order and with the possibility of finishing early.