Changed "split" algorithm to go from split node up to root
This changes the algorithm for "split" to accumulate pre- and post-split trees by moving from the split node up to the root of the tree, rather than from the root down to the split node.

This is less code, and it's more performant. In particular, it reduces the number of calls to augment(). In the downward approach, augment() could be called multiple times on a given node. This is because calls to fixInsertionWithoutGettingRoot might augment some nodes during rotations, but some of these augmentations would be invalidated as nodes were added to the tree below the augmented nodes. At the end of the "split" method, we would unconditionally call augment() on all of the nodes on the spines of the resulting trees, in order to fix the augmentations after the fact.

By contrast, the upward approach is conducive to doing a minimum number of augmentations. There is no risk of calling augment() on a node, and then later altering its descendants and having to call augment() on the node again, because we only ever progress upward. Furthermore, whenever augment() returns false, we can avoid unnecessary further calls to augment().